Rework uses_shed...
